MyApplication/app/src/main/java/com/example/myapplication/ui/main/viewmodel/MainViewModel.kt
package com.example.myapplication.ui.main.viewmodel

import androidx.lifecycle.*
import com.example.myapplication.data.api.ApiHelper
import com.example.myapplication.data.model.User
import com.example.myapplication.utils.Resource
import kotlinx.coroutines.Dispatchers
import kotlinx.coroutines.launch
import java.lang.Exception

class MainViewModel(private val apiHelper: ApiHelper) : ViewModel() {

    fun getUsers() = liveData(Dispatchers.IO) {
        emit(Resource.loading(data = null))
        try {
            emit(Resource.success(data = apiHelper.getUsers()))
        } catch (exception: Exception) {
            emit(Resource.error(data = null, message = exception.message ?: "Error Occurred!"))
        }
    }
}